Extract common TEventProcessorRunner::run
