+++ nexus-watcher/src/service/runner/mod.rs
@@ -5,9 +5,11 @@ pub use homeserver::HsEventProcessorRunner;
 pub use key_based::KeyBasedEventProcessorRunner;
 
 use std::sync::Arc;
+use std::time::{Duration, Instant};
 
 use nexus_common::types::DynError;
 use tokio::sync::watch::Receiver;
+use tracing::{error, info};
 
 use crate::service::{
     indexer::{RunError, TEventProcessor},
@@ -51,14 +53,58 @@ pub trait TEventProcessorRunner: Send + Sync {
     /// Determines the list of target IDs to process in this run cycle.
     async fn pre_run(&self) -> Result<Vec<String>, DynError>;
 
-    /// Main run loop: builds and runs event processors for the relevant targets.
-    ///
-    /// # Returns
-    /// Statistics about the event processor run results, summarized as [`ProcessedStats`]
-    async fn run(&self) -> Result<ProcessedStats, DynError>;
-
     /// Post-processing of the run results.
     ///
     /// Receives the raw stats from the run and returns processed stats.
     async fn post_run(&self, stats: RunAllProcessorsStats) -> ProcessedStats;
+
+    /// Main run loop: builds and runs event processors for the relevant targets.
+    ///
+    /// # Returns
+    /// Statistics about the event processor run results, summarized as [`ProcessedStats`]
+    async fn run(&self) -> Result<ProcessedStats, DynError> {
+        let hs_ids = self.pre_run().await?;
+        let mut run_stats = RunAllProcessorsStats::default();
+
+        for hs_id in hs_ids {
+            if *self.shutdown_rx().borrow() {
+                info!(hs_id = %hs_id, "Shutdown detected; exiting run loop");
+                break;
+            }
+
+            if let Some(skip_status) = self.backoff_should_skip(&hs_id).await {
+                run_stats.add_run_result(hs_id, Duration::ZERO, skip_status);
+                continue;
+            }
+
+            let t0 = Instant::now();
+            let status = match self.build(hs_id.clone()).await {
+                Ok(event_processor) => status_from_run_result(event_processor.run().await),
+                Err(e) => {
+                    error!(hs_id = %hs_id, error = %e, "Failed to build event processor");
+                    ProcessorRunStatus::FailedToBuild
+                }
+            };
+            let duration = t0.elapsed();
+
+            self.backoff_on_result(&hs_id, &status).await;
+            run_stats.add_run_result(hs_id, duration, status);
+        }
+
+        let processed_stats = self.post_run(run_stats).await;
+        Ok(processed_stats)
+    }
+
+    /// Called before processing a homeserver, to check if backoff mechanism indicates it
+    /// should be skipped. Return `Some(status)` to skip it.
+    ///
+    /// No-op default implementation. Runners that use backoff should overwrite as needed.
+    async fn backoff_should_skip(&self, _hs_id: &str) -> Option<ProcessorRunStatus> {
+        None
+    }
+
+    /// Called after a homeserver is processed (build + run), to update its backoff status.
+    ///
+    /// No-op default implementation. Runners that use backoff should overwrite as needed.
+    async fn backoff_on_result(&self, _hs_id: &str, _status: &ProcessorRunStatus) {}
 }
